requirements.txt
accelerate
transformers
peft
datasets
bitsandbytes
torch
pillow
gradio
opencv-python-headless
python-dotenv
pyyaml
